カウントダウンの修正が完了した、リザルトシーンのブランチなのに...... 画像をレイアウト保存できた、あとは差し替えていい感じにする

+++ Engine/Global.cpp
@@ -47,12 +47,13 @@ XMFLOAT3 operator-(XMFLOAT3& a, const XMFLOAT3& b)
 
 // ゲームのスコア
 namespace ScoreManager {
-	int g_Score = 0;
+	float g_Score = 0.f;
 
 	void ScoreManager::AddScore(int score) {
 		g_Score += score; 
 	}
 
-	int ScoreManager::GetScore()
-	{ return g_Score; }
+	int ScoreManager::GetScore(){
+		return g_Score; 
+	}
 }
